requirements.txt
asgiref==3.6.0
autopep8==2.0.1
cffi==1.15.1
cryptography==39.0.2
Django==4.1.6
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
